% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AhoCorasickTrie.R
\docType{package}
\name{AhoCorasickTrie}
\alias{AhoCorasickTrie}
\alias{AhoCorasickTrie-package}
\title{AhoCorasickTrie: fast searching for multiple keywords in multiple texts}
\description{
AhoCorasickTrie: fast searching for multiple keywords in multiple texts
}

